LayoutTests/http/tests/security/contentSecurityPolicy/audio-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load http://127.0.0.1:8000/resources/redirect.py?code=307&url=http%3A%2F%2Flocalhost%3A8000/resources/balls-of-the-orient.aif because it does not appear in the media-src directive of the Content Security Policy.
ALERT: PASS
